refactor: Inline `usePortfolioURLEasterEgg` hook to ProjectCard
